requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
asyncpg
pymysql